Fix add/del user to group trigger USERGROUP_MODIFY
Adding or removing a user from a group modifies the user object on Dolibarr's side.
In LDAP however, members of a group are stored in the group itself.
Therefore group must be updated after adding/removing a user from it.
Trigger the USERGROUP_MODIFY event from SetInGroup/RemoveFromGroup to that effect.